Split the code in multiple files
