+++ Pipfile
@@ -7,9 +7,13 @@ name = "pypi"
 numpy = "*"
 matplotlib = "*"
 scipy = "*"
-
+isort = "*"
+black = "*"
+flake8 = "*"
 
 [dev-packages]
+pytest = "~=6.2.5"
+"pytest-flake8" = "~=1.0.7"
 scikit-learn = "*"
 
 [requires]
